Printing results enabled

+++ preload.js
@@ -1,6 +1,8 @@
 
 const db = require('./dbFunctions.js');
 
+
+
 function selectResult(result){
   console.log(result);
   console.log(result.rowsAffected[0]);
@@ -11,6 +13,14 @@ function selectResult(result){
 //success is a callback function.
 function success(result){
   console.log(result);
+  var html = '';
+
+    setTimeout(function () {
+
+            html += '<p>' + result + '</p>'
+
+        document.getElementById("success-result").innerHTML = html
+    }, 500)
 }
 
 
@@ -23,7 +33,12 @@ function insertForm(event) {
     add2 = document.getElementById("add2").value;
     district = document.getElementById("district").value;
     phone = document.getElementById("phone").value;
-    console.log(district);
+    console.log(name);
+    if(name=='' || phone =='')
+    {
+      success("Name & Phone no. are mandatory");
+      return;
+    }
     db.insertData(name,add1,add2,district,phone,success);
     document.getElementById("form").reset();
 }
@@ -36,6 +51,11 @@ function updateForm(event) {
     add2 = document.getElementById("add2").value;
     district = document.getElementById("district").value;
     phone = document.getElementById("phone").value;
+    if(name=='' || phone =='')
+    {
+      success("Name & Phone no. are mandatory");
+      return;
+    }
     db.updateData(name,add1,add2,district,phone,success);
     document.getElementById("form").reset();
 }
@@ -45,6 +65,12 @@ function deleteForm(event) {
     console.log("CALLED");
     name = document.getElementById("name").value;
     phone = document.getElementById("phone").value;
+    if(name=='' || phone =='')
+    {
+      success("Name & Phone no. are mandatory");
+      return;
+    }
+        console.log(name);
     db.deleteData(name,phone,success);
     document.getElementById("form").reset();
 }
@@ -68,11 +94,23 @@ function viewData()
   db.displayData("",viewDataCallback);
 }
 
+function printData()
+{
+  event.preventDefault() // stop the form from add1itting
+  document.getElementById("body").style.display = "none";
+  document.getElementById("printBtn").style.display = "none";
+  window.print();
+  document.getElementById("body").style.display = "block";
+  document.getElementById("printBtn").style.display = "block";
+}
+
+
 function viewSearchData()
 {
   event.preventDefault() // stop the form from add1itting
   name = document.getElementById("name").value;
   phone = document.getElementById("phone").value;
+
   console.log(name);
   var whereClause = `where name='${name}' or phone='${phone}'`; //+ 'or phone=' + phone;
   db.displayData(whereClause,viewDataCallback);
